+++ packages/uba/uba_usercertificate.js
@@ -21,7 +21,11 @@ function setBlob (ctxt) {
   const execParams = ctxt.mParams.execParams
   Object.keys(execParams)
   if (execParams.certificate) {
-    execParams.setBLOBValue('certificate', Buffer.from(execParams.certificate, 'base64'))
+    const iitCrypto = require('@ub-d/iit-crypto')
+    const certBin = Buffer.from(execParams.certificate, 'base64')
+    const certJson = iitCrypto.parseCertificate(certBin)
+    execParams.certParsed = JSON.stringify(certJson)
+    execParams.setBLOBValue('certificate', certBin)
   }
 }
 
@@ -55,7 +59,7 @@ function clearBlobAndLogoutAUser (ctxt) {
  * @private
  * @param {ubMethodParams} ctxt
  */
-function clearBlob(ctxt) {
+function clearBlob (ctxt) {
   const execParams = ctxt.mParams.execParams
   if (execParams.certificate) {
     execParams.certificate = ''
@@ -63,22 +67,42 @@ function clearBlob(ctxt) {
 }
 
 /**
- * Retrieve certificate as base64 encoded string
+ * Retrieve certificate as:
+ *  - base64 encoded string, if called as ublq
+ *  - binary, if called as `/rest/uba_usercertificate/getCertificate?ID=223`
  *
- * @param {ubMethodParams} ctxt
- * @param {number} ctxt.mParams.ID
+ * @param {ubMethodParams} [ctxt]
+ * @param {number} [ctxt.mParams.ID]
+ * @param {THTTPRequest} [req]
+ * @param {THTTPResponse} [resp]
  * @method getCertificate
  * @memberOf uba_usercertificate_ns.prototype
  * @memberOfModule @unitybase/uba
  * @published
  */
-me.getCertificate = function (ctxt) {
+me.getCertificate = function (ctxt, req, resp) {
+  let certID
+  if (req) { // endpoint is called as rest/uba_usercertificate/getCertificate?ID=1231
+    if (!req.parsedParameters.ID) {
+      return resp.badRequest('Missed ID; Expect URL to be rest/uba_usercertificate/getCertificate?ID=1231')
+    }
+    certID = req.parsedParameters.ID
+  } else {
+    certID = ctxt.mParams.ID
+  }
   const store = UB.Repository('uba_usercertificate')
     .attrs(['ID', 'certificate'])
-    .where('ID', '=', ctxt.mParams.ID).select()
+    .where('ID', '=', certID).select()
 
+  if (store.eof) throw new Error('not found')
   let certificate = store.getAsBuffer('certificate')
-  certificate = Buffer.from(certificate)
-  certificate = certificate.toString('base64')
-  ctxt.dataStore.initFromJSON({ fieldCount: 1, values: ['certificate', certificate], rowCount: 1 })
+  if (req) {
+    resp.writeEnd(certificate)
+    resp.writeHead('Content-Type: application/x-x509-user-cert')
+    resp.statusCode = 200
+  } else {
+    certificate = Buffer.from(certificate)
+    certificate = certificate.toString('base64')
+    ctxt.dataStore.initialize({ fieldCount: 1, values: ['certificate', certificate], rowCount: 1 })
+  }
 }